Add source group (i.e. fallback source) support
Source distributions can now be put in groups, to allow e.g. updating
from precise-updates, and if a package is not available in
precise-updates, automatically fall back to precise.

To make dealing with source groups easier, and to prevent the abuse of
the concept of "component" for OBS-based targets (which e.g. made it
impossible to generate merge stats for multiple OBS target distros in
one MoM installation), target distributions and their default and
optional source groups are now explicitly defined in momsettings in
the DISTRO_TARGETS structure. MoM subcommands that formerly operated
on a component, or on a specific target distro/suite, now operate on
a distribution target (via the --target switch).