Ignore non-newest pool versions for most operations
It's possible for multiple revisions of a package to exist in an obs
distro's mom pool. If we do not restrict ourselves to the newest, we
risk pathological behaviors, e.g. repeatedly regenerating the newest
revision of a package from the oldest by merging the oldest with a
source distro version, and then repeatedly committing that regerated
version to obs.