commit_merges: when updating to an upstream version, use that version
package is a Package in target.distro, so package.newestVersion()
is the newest version *in the target distribution*. If we're updating
an source package that has no local changes in the target distribution,
but has been updated in the source distribution, then we will have to
use the source distribution's version number when looking for the
right pool directory from which to copy the sources.

(The merge process already needed to look at those sources, so ideally
the report would tell us where to get them, but this will do for
now.)

This commit only changes the "commit to a branch and submit a request"
workflow (committable=False). The "commit directly to target
distribution" workflow (committable=True) appears to rely on us having
copied the updated sources into osc/DISTRO/PROJECT/PACKAGE/ already;
it isn't clear to me whether that actually happens correctly or not.

Bug: Collabora #389
Reviewed-by: Sjoerd Simons