fetchMissingVersion: only fetch what we actually want
Previously, for each suite (aka dist, release, codename; e.g.
"unstable"), we asked the Distro object (e.g. debian) for every
version of the given package in any suite, and tried to download
them all. This would fail with an unhandled exception if there is
any suite that we can't search because we haven't downloaded
Sources.gz - for instance, if Debian (stable, testing) are configured
in DISTRO_SOURCES, but everything in DISTRO_TARGETS is currently only
using stable, we won't have downloaded testing's Sources.gz.

Now we only search the sources configured in DISTRO_TARGETS,
and we only look for the version we actually want.

Bug: Collabora #29
Reviewed-by: Sjoerd Simons