fetchMissingVersion: cope with PackageNotFound
In a more complicated source list, it's not necessarily the case that
every package exists in every suite ("dist"). For instance,
wheezy-updates isn't self-contained, and neither is security.debian.org.

This is not an exceptional condition, so suppress the log message
when not in verbose mode. Arguably, we should make more noise if
we can't find the desired package/version *anywhere*... but in practice
we don't download missing versions unless we want to use them for
something, so the unpack will fail and the failure will be logged.

Bug: Collabora #29
Reviewed-by: Sjoerd Simons