produce_merges: read changelogs to find the common ancestor
We previously used target.findNearestVersion, but that function
doesn't have enough anywhere near enough information to find a
correct base version, and should be deprecated or deleted.

One particularly obvious example of where it failed: if Debian has
gone from 1.2-3 to 1.2-4 to make change A, and our derived distro has
gone from 1.2-3 to 1.2-3deriv1 to make change B, we would consider
1.2-3deriv1 to be the base, and conclude that we should update to
1.2-4 for change A, discarding change B.

The algorithm used in findNearestVersion does hard-code that the
suffix "co" + digits should be discarded when searching for a base, so
replacing -3deriv1 with -3co1 would work correctly; but this doesn't
work in more complex cases. For instance, if Debian updates from
1.2-3 to 2.0-1 (perhaps with fixes beyond just bringing in the new
upstream version), and we have updated from 1.2-3 to 2.0-0co1
(again, perhaps with further fixes), the correct answer is a
three-way merge to get the union of both sets of changes.

Bug: Collabora#29
Reviewed-by: Sjoerd Simons