produce_merges.copy_in: take a PackageVersion
There's little point in taking a Distro and a Sources.gz stanza, then
searching that Distro for a matching package, if we could just as
easily take a PackageVersion and use that.

Also, the findPackage can fail with an exception if we don't have a
cached Sources.gz for one of its suites (if we have configuration
for stable, testing and unstable, but all targets pull from either
stable or unstable, then debian-testing/main/source/Sources.gz will
be missing), or pick the wrong pool directory (like Collabora#389).

Bug: Collabora#390
Reviewed-by: Sjoerd Simons