Read and write merge reports using JSON
Strictly speaking, JSON doesn't have comments, so I'm using the
convention that a key named "#foo" is a comment about "foo", or a
human-readable version of "foo".

This does introduce one regression, fixed later in this branch: at the
moment we only write out the committed:True marker to the non-JSON
version, and we prefer to read the JSON version, which means we repeat
the merge in every run, even when the versions have not changed since
the version we committed (presumably to a branch where it is awaiting
review).

Bug: Collabora#392