produce_merges: look for common ancestor in target, not only sources
Previously, we'd only look for the common ancestor in "their" distros,
not in "our" distro. This works fine if we're tracking Debian stable
(which keeps superseded packages referenced), but fails if we're
tracking Debian unstable or Ubuntu (which don't generally do that)
and our package is an older version taken from Debian/Ubuntu.

Bug: Collabora #418