Be more careful with findPackage returning a list
findPackage returns a list of PackageVersion instances in arbitrary
order, so we can't rely on the first one in the list being the newest.
In particular, when a Debian stable point updates a package to a newer
version, the obsoleted version is not removed: if the older version
happens to be the first one listed in Sources.gz, we won't notice
the newer version for merge purposes.

Bug: Collabora#420