OBSDistro: merge sync, update and checkout into one "sync" method
checkout() appears to have regressed when it was changed to take
Package arguments in commit 2da9871: str.join can't act on a sequence
containing an OBSPackage. However, the exception was ignored in its
only caller, sync(), until 2f37806, and update() fixed the situation
afterwards.

The only one of these functions that's called externally is sync(),
and the only part of sync() that works is update(), so just rename
update() to sync() and delete the other two.