PoolDirectory.updateSources: don't second-guess whether we must update
Comparing the mtime of each file with the mtime of the directory
is OK up to a point, but it can't catch files being deleted.
We could use the mtime of the directory, but only because we don't
use a copy + atomic-rename-to-overwrite process for the Sources file,
which itself is more like a bug than a feature; if we started using
atomic updates, we'd update the directory's mtime every time we did
that, so it'd always be newer than Sources and we'd always update it.

Instead, expect callers to track whether things have changed, and just
do the update when asked.

Bug: Collabora #430