merge_report: document merged_patch correctly
I thought it was diff(left..proposed), i.e. the change that will happen
in our distro if the proposed merge is accepted.

It's actually diff(right..proposed), i.e. the delta that will exist
from their (upstream) distro to our (downstream) distribution
if the proposed merge is accepted. That's the right thing to review
if you're writing a summary of changes relative to the parent distro,
but not necessarily the right thing to review if you're checking
whether the change is OK to accept.