Check for base versions in pool directories, not just on upstream servers
If the base version has disappeared from the upstream server, we
might still have a copy locally - either because it used to be current,
or because someone explicitly downloaded it in response to the advice
given in the emails.

Rather than partially reverting commit 3a160af, I'm just updating
Sources unconditionally - it's rather faster than the actual unpack
and merge, which we're about to do anyway.

Bug: Collabora #443