Allow preferring upstream's version without merging
Add a new configuration key, sync_upstream_packages, that allows setting
a list of packages where the upstream version is preferred without
trying to reapply changes.

The use case is rebasing a distro from one upstream to another. In some
cases, the changes that were being applied may not be interesting or
needed and just represent an unnecessary delta from the new upstream.
While this can be managed by declining MOM's merge request and doing the
update manually, that's very tedious for a full distro rebase.

https://phabricator.endlessm.com/10577