OBSDistro.sync: Expand OBS links like osc checkout does
Without this, there will just be a _link file in the checkout. This
breaks when the linked version in an existing checkout is updated. m-o-m
considers only the files in the checkout, but those might be stale if
the link isn't expanded.

Expand new checkouts like "osc checkout" does by default. For existing
checkouts, find the linked revision and update to that like "osc up -e"
does.

https://phabricator.endlessm.com/T10577