produce_merges: Sync to upstream if changelog is only diff
If the only difference between the upstream and merged versions is in
debian/changelog, then we can drop that noise and simply use the
upstream version.

The process used here is pretty hacky as a throwaway patch is used and
then analyzed with diffstat. It would be better if do_merge() returned a
list of changed files, but that would be more involved.

https://phabricator.endlessm.com/T10889