OBSDistro.sync: Try to unfreeze packages when updating
In certain scenarios with branched packages, osc will "freeze" the link
to the target package. This primarily seems to occur when the branched
package is updated, then the target package is updated, and updating
would cause a conflict between the two. Once the link is frozen, updates
will no longer pull in changes from the target package and the commit
will be made incorrectly since it doesn't remove the original source
files.

While "osc pull" has some complicated code to do a 3-way merge, it seems
that we should just be able to unfreeze the link and update since m-o-m
is simply going to throw out any files there and add the new merged
files.

https://phabricator.endlessm.com/T16751