commit_merges: Remove linked files, too
Since 276ed19, the OBS checkouts have been expanded so that the linked
sources are downloaded, too. This was so that if the linked target
project was updated, m-o-m would remove the new base source files and
the commit would show up properly as an add and remove.

Unfortunately, m-o-m wasn't actually deleting the linked sources because
it was only considering the branched sources for deletion. Query the
linked target package file list and delete those, too.

https://phabricator.endlessm.com/T16751