Remove PoolDirectory and rework Sources parsing
PoolDirectory was a class that abstracts a pool for a specific
package, and it maintained a Sources file describing the contents
generated using apt-ftparchive.

The excessive running of apt-ftparchive (especially in the
produce_merges path) causes a slowdown, and in general this
seemed like an unneeded abstraction. All the info we need is
available in the .dsc files so we can just read them directly.

When converting PoolDirectory users to use something else, I also
ended up reworking many instances of package discovery in general.
The codebase has classes that abstract packages and versions, but
much of the code was using Sources file parsing as a basis.

After these changes, the Distro, Package and PackageVersion classes
become the main way of iterating through the work that needs to be
done.

I tried to keep changes somewhat minimal, but get_common_ancestor()
was quite a headache to understand, so I ended up reworking it a
bit. This part of the code will be revisited again in future.