produce_merges: save conflicts at end of merge
All of the codepaths that need to record that a file can't be merged
additionally directly copy the conflicting versions to the merge output
directory.

Simplify the code and allow for some further reworks by having
the multiple merge codepaths just record which files are conflicting,
deferring the conflicting version copy procedure to the end of the
merge.