produce_merges: Return information from do_merge
Currently, do_merge only returns a list of conflicting files.

Extend this to return information about the changes that were made
relative to the new version that we are merging against.

The code that forces a sync if the changelog was the only change
has been cleaned up to use this information.