Rework detection of base and upstream versions
Previously there was a confusing mess regarding which versions are
considered as base and upstream. update_sources made some attempt to
identify base and upstream versions and download them, then
get_missing_bases had a slightly different approach to finding the base,
then produce_merges had a slightly different approach and would even
step out and try to download more things in some cases.

Sanitize all this into the update_sources step. We now identify and
download the base and upstream versions all in that place, and we
record the results in a new UpdateInfo file which is then used in
the later steps in order to rely on the same decision of which is the
base and upstream to use.

The steps taken now should be clearer and consistent, and should also
be more intelligent when finding bases. There was prevously an inclination
to finding a common ancestor, but that was leading to unnecessary merges,
for example the upgrade from patch_2.7.5-1+deb8u1 to 2.7.5-1+deb9u1 was
being done as a merge instead of a sync.

We now detect when our version is unmodified (based on version number)
and resort to syncs in such cases. When we need to find a base version,
we consider any previous unmodified version, which does not need to be a
common ancestor. This should result in more appropriate base versions
being found and used.