Add merge report notes
Allow the merge process to flag up specific notes which are then
shown in the OBS merge request message and also on the report
website.

To start with, just add a single message to specifically flag
when we reverted to the upstream version just based on changelog
modifications. More messages will be added in later changes.