DebTreeMerger: record and apply merge changes separately
When merging, when a downstream change is detected, do not merge
it immediately. Instead just add it to a pending_changes list.

This allows for an intelligent context-specific merge to take place
before the "generic" code is then executed on all pending changes.
This cleans up the mess around .po file handling and will allow for other
context-specific merge code to run here too.

Also, rather than three sets, consolidate the changes that were made in the
merge into a single data structure to organise things consistently.