DebTreeMerger: support basic conflict resolution of quilt series files
If we were unable to merge the series file using the usual methods,
and if our downstream changes were just appending new patches to the
end of the list, resolve the conflict by readding our additions
to the end of the new right side series file.

This fixes the update of cairocffi_0.5.4-1endless1 on top of new upstream
version 0.7.2-2.