Handle multiple upstream sources as a sequential list
With Endless currently based on Stretch, we also have some packages
that come from stretch-backports and some that come from buster.

Currently merge-o-matic wants to upgrade the stretch-backports packages
to buster (the distro configured as unstable), but that's usually not what
we want.

Tweak the handling of upstream sources so that they are treated as a
list of distros. Only the first distro that provides a better version
is considered. This will allow us to add stretch-backports as an earlier
unstable upstream, and the packages that originate from there won't
be automatically pushed up towards buster. One such current example
is debhelper.