produce_merges: find upstream version via Sources
Using target.getAllPoolVersions() to find the upstream version is imperfect.

The pool thing may need a bit of a rethink, because pools are shared then
Package.getPoolVersions() can return a PackageVersion that appears
to belong to a given distro even though that is not the case.

To avoid this, find the upstream version using findPackage(), which
consults Sources files and hence will only return packages that
really belong to the right distro.

This fixes the merge reports incorrectly showing the wrong source
of the new upstream version.

This is not needed for the base version search, where it's probable
that the base version is no longer in any current distro, and we really
want to do a "deeper" search in our pools to find it.

I might rework or at least better document the caveats around
getPoolVersions() later.