Reset debian/control Uploaders field
Some downstream development processes end up making modifications
to the Uploaders field, such as in the case of cheese_3.26.0-1endless3.
Such changes tend to cause conflicts, which is true in the case of
the upgrade to cheese_3.28.0-1.

Add logic to detect Uploaders changes and drop those changes, to see
if that helps avoid merge conflicts.

The ControlFilePatcher class will probably be extended later to deal
with other cases where we can intelligently resolve control changes.