Integrate State-Based Text Merge tool
This tool is cleverer than diff3 in terms of merging.
It crashes due to heavy recursion in many cases, but in some cases
it works fine.

For example upgrading colord_1.2.1-endless1 to color_1.3.3-2,
diff3 finds conflicts in debian/rules but sbtm is able to produce a
correct merge (I used this as a basis for the test case).

When it fails, it's usually because it exhausts recursion limits
in the diff algorithm, so we add a fallback on the alternative fast
algorithm. This at least produces a correct merge for debian/control
in the same colord upgrade.