Add --use-upstream parameter to force a specific upstream
This is useful in the case of e.g. libcanberra, where it has been
determined that the current proposed update to the latest stretch
version is not going to work for us. We want to switch to the
buster version instead.

Add a command line parameter to make merge-o-matic try the merge with
a distro of your choice.

Adjust the update_sources logic to only proceed towards merging
if the upstream version is newer than any previous UpdateInfo
that was saved. This means that if you run as a one-off to
upgrade libcanberra to buster, then merge-o-matic runs with
default parameters which would sync it to stretch, it would
honour the existing buster merge since the upstream version
is newer.