DebTreeMerger: detect and drop downstream patches which can be reverted
If a downstream patch can be reverted from the new version then
it implies that the patch was accepted upstream.

Also support chains of patches that can be reverted, and handle the case
when there are no patches left (so debian/patches should be removed).

This allows us to automate the upgrade of geoclue-2.0 from 2.4.7-1endless1
to 2.4.12-2. The downstream patch 03-bfdo-102519-T18994.patch went upstream
and can be cleanly reverted.