Add dependency parsing and merging
Extend the parser to understand Depends/Build-Depends fields.

Catch the case where a downstream package modifies the version
constraint of a build dependency, where the new upstream version
has dropped that build dependency altogether. In this case,
our local change can be dropped.

This enables the automatic merge of appstream_0.10.6-2endless1 to
appstream_0.12.2-2, and helps with the merge of
fonts-cantarell_0.0.25-2endless1 to 0.111-2.