DebTreeMerger: fix control file merge tracking
One of the tests was wrong here - after reverting the Uploaders
change (the only downstream change), the control file should not
be regarded as being modified relative to the right version.

Clean up the logic so that handle_control_file() doesnt actually
have to perform the merge (and do book-keeping), instead just
check for mergeability. Then modify the generic codepath to
realise when no merge is ultimately needed.