debtreemerger: detect duplicate patches in left and right
When upgrading cogl_1.22.2-2endless1 to cogl_1.22.2-5, the
Endless patch we have can be reverted on the basis that Debian
has added an identical patch.

Detect this by adding another patch reversion check that is run
only after the right side patches have been applied, automating
the merge of this cogl upgrade.