Attempt patch refresh via sbtm
If we can't apply a patch with fuzz to refresh it, fall back on
recreating the precise base/left/right versions and using sbtm
to see if it can figure out the merge. If so, recreate the
patch based on sbtm's output.

This fixes the update of dracut_044+241-3endless2 to 048+80-1.