Merge debian/control.in too
Use the existing parser/merger. This helps with the merge of many packages.

Also address the case where control.in is dropped on the right,
such as in the update from libffado 2.3.0-2endless2 to 2.4.1-0.1.
In this case, we can assume that the relevant changes are also present in
the control file, so it is safe to drop our downstream changes.