debtreemerger: fix patch reversion
"patch -R --force" will actually sometimes revert to applying the patch
in forwards (non-reversed) mode, and return success. That's not what
we were looking to allow here.

Instead, redirect patch stdout to /dev/null, this will stop it from
asking questions and it will not fall back on applying patches
in forwards mode.

Also note that the series file was modified in this codepath,
to ensure correct file state tracking.