debtreemerger: handle convergance of changes into right version
Seen with the upgrade from vino_3.22.0-1endless1 to 3.22.0-4,
we got into a state where the merged version of debian/rules
was identical to the right version, however we had not detected
the convergence due to unrelated changes made on the right, and
were reporting that debian/rules still had local changes.

Use a md5sum check to catch this.